Fix crash that occurs between 15 to 20 rounds of play.
When the server throws the error "ActionDispatch::Cookies::CookieOverflow", the display of activities in the message window stops scrolling. Play may continue with bad display in message window, or application may crash. Problem occurs because the messages are being stored in session and the size allocated for session is exceeded.

The fix limits the number of saved messages by deleting the oldest message when adding a new message exceeds the count of allowed messages.